requirements.txt
pandas
numpy
yfinance
loguru
pyyaml
pydantic
pydantic-settings
SQLAlchemy>=2.0
plotly
kaleido
